api/requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
gunicorn==19.9.0
python-dotenv==0.10.3
